Add facility to use as a submodule
I don't really want to simply copy the code.  I want to keep it under
its own version control so I can pass my updates around.  To do this, we
need to be able to install as a submodule.  Luckily, consulting the
_Configuration File Reference_ portion of the scons(1) man page tells me
we can do this using a module.  However, we need to package the tool
with an `__init__.py`.  To make it work, we just import everything from
the main tool module.