Update the tree walking logic
Ultimately, we want to drill down to the lowest leaf and flatten
everything into a single list.  This means we need to flatten the list
as we go so that each return is working on a single list.  We can also
clean up the code and make it a bit more Pythonic in the list
comprehension and checking.