Process the document through each filter correctly
To get the final AST, we need to run each JSON filter passing the format
as an argument to the filter.  This means we need to split the command
line into components and run each via pipes as described in the Pandoc
User's Guide.  We need to make sure the filters are run in the
appropriate order and the Lua filters are properly sequenced.