Only scan for images in a final output
If the generated document still needs processing, we have no need to
scan for image dependencies.  For example, if we are generating a LaTeX
file that will be processed by SCons, we don't want to waste time
looking for Images or `\input`s.  SCons already does that and does it
better.  Likewise, if we use a Markdown file as an intermediary for
converting to docx, we only care about the images needed to build the
docx not the Markdown.