Move the from flag to the front of the command
We need the from command at the start of the command line so we know
what the user expects to tell Pandoc to parse.  If we don't, we'll use a
default and that is not necessarily right.  Further, the from flag
latter in the processing will break the filter piping.