Carefully handle the first process call
We cannot simply call Pandoc with the sources and then process the rest
of the command.  What happens if the user provided a --from?  The answer
is a crash.  What we really want to do is process the command line up to
the first filter and then add the sources using the JSON output.  We
need to do this if we have a filter or not.