Fix the regex for the filter
I forgot to take into account a possible '-' in a filter name.  The
pattern would match up to a '-', and that is wrong.  I _think_ I've
caught all use cases.  The only one I can think of is when the filter
has a space in the name, but that's a bad idea under any case.