Account for directories when grabbing filter
The user can specify a relative path to the filter.  This means we need
to account for '/' and '.' in the filter name.  Otherwise, we don't get
the correct filter.  Note this could create a dependency outside of the
build!  This is generally not a good thing, but I'm not sure how to
prevent this in a robust way.