Add a test session to make sure it all works
The end goal is to make sure the tool just works.  Instead of hand
running the example, we can add a test session to the nox file.  It just
makes sure the generated HTML is what is expected.  We cannot check the
graphic because matplotlib drifts too much.

And while we're at it, we stop ignoring the reference example files but
ignore the .nox directory.