Add the support files for setuptools
To distribute the project, we need some additional tools.  However, in
order to follow the "copy this directory to site_scons/site_tools", we
need an __init__.py at the top of the directory tree.  It's sole job is
to import all the appropriate symbols into the local namespace.

In the process, we update the panflute version in the noxfile.py.