Update the installation instructions
It turns out scons restructured its installation with version 4.  Prior
to that, the fixed up installation does not work.  For scons prior to 4,
the user will have to copy the directory to the appropriate place in the
project or .scons directory.  That also means the test needs to be
updated to copy as appropriate.