Reduce Pandoc+panflute version restrictions
Pandoc pre 2.10 still works with panflute pre 2.0 so we should allow it.
This means we need to check the versions of Pandoc and panflute to
ensure compatibility early and not wait for an obscure error which could
make little sense to the end user.  We also need to update the test and
handle the fact that earlier Pandoc versions generated slightly
different HTML output.